requirements.txt
numpy
scikit-learn==1.4.2
xgboost
lightgbm
catboost
umap-learn
tensorflow==2.16.1
torch==2.3.0
fuzzywuzzy
python-Levenshtein
PyYAML
fastavro
pyarrow
# llm
langchain-openai
python-dotenv
langchain-chroma
langchain-community
certifi
httpx